fix(floating-btn/settings): Resolve issues after splitting `settingsManager`
- Address and resolve errors introduced by splitting `settingsManager`
- Ensure settings functionality is restored and working correctly

Changes summary:
- Fixed the errors and issues that arose from splitting the `settingsManager`, restoring the settings functionality to its previous working state.
